Add `Certifiates.SANsEntriesCount` field
This field would allow the payload creator to omit SANs entries
to conserve plugin output size and still indicate the number of
SANs entries present for a certificate. This helps prevent a
scenario where an empty SANs entries collection implies that no
SANs entries were provided (though the
`CertificateChainIssues.MissingSANsEntries` field if checked
would dispel that notion) but also provides the number for use
in display or for metrics purposes.

refs GH-33