Add Certificates collection type & initial methods
Add helper methods/functions for comparing and displaying
`Certificate` values from a `Certificates` collection.

As part of this, the certificate type constants within the
internal/certs package have been exported for wider use
within this library (not externally).

This new `Certificates` collection and behavior can be used
when generating reports.

NOTE: I'm on the fence whether providing this support fits
within the scope of this project. The intent is to generate
and deconstruct certificate metadata payloads and let consumers
of that metadata handle grouping, evaluating and displaying the
data in formats of their choosing.

By providing even a little support to analyze and display the
data this library takes on a role it was not *exactly* intended
to handle.

In short, this support may be pulled later.