Fix handling of leaf certificates in collection
Earlier logic failed to take into consideration that leaf cert type
includes self-signed leaf certs also. Update logic to look for either
leaf certificate type.

Fix misreport of missing cert chain for bundle.

It's OK for a leaf cert not to be present in an intermediates only
cert bundle, provided there is something in the Certificates
collection.

If we have something, report just the cert being evaluated as missing,
otherwise we then report the whole chain as missing.