Adjust misordered chain logic
Switch from considering all signature verification errors as
a confirmed identification of a misordered cert chain to
explicitly looking for `x509.InsecureAlgorithmError` and ignoring
those. In that scenario we rely exclusively on the issuer/subject
equality check.

If any other signature verification error occurs we consider
that scenario an indicator of a misordered cert chain.

No signature verification error for current/next certificate is
interpreted as the certificates being in the correct order.

refs GH-72